% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{dat_hospital}
\alias{dat_hospital}
\alias{dat_ipd}
\alias{dat_icu}
\title{Example hospital datasets containing proper names in different formats}
\format{
Data frames each with two columns:

\describe{
\item{name_ipd/name_icu}{patient name}
\item{date_ipd/date_icu}{date of entry to given department}
}
}
\description{
Example hospital datasets, one from an in-patient department
(\code{dat_ipd}), and the other from an ICU department (\code{dat_icu}). The datasets
contain some common patients but with variation in how the names are written.
Note these data are fake -- the patient names are simply random combinations
of common French names.
}
