% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_eval.R
\name{match_eval}
\alias{match_eval}
\title{Evaluate token match details to determine overall match status}
\usage{
match_eval(k_x, k_y, n_match, n_match_crit, ...)
}
\arguments{
\item{k_x}{Integer vector specifying number of tokens in names \code{x}}

\item{k_y}{Integer vector specifying number of tokens in names \code{y}}

\item{n_match}{Integer vector specifying number of aligned tokens between x
and y that are matching (i.e. based on argument \code{dist_max} in
\code{\link{nmatch}})}

\item{n_match_crit}{Minimum number of matching tokens for names x and y to be
considered an overall match}

\item{...}{Additional arguments (not used)}
}
\value{
Logical vector indicating whether names \code{x} and \code{y} match, based on the token
match details provided as arguments
}
\description{
Evaluate token match details to determine overall match status
}
