% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_eval_token.R
\name{match_eval_token}
\alias{match_eval_token}
\title{Evaluate string distance and token lengths to determine whether two tokens
match}
\usage{
match_eval_token(nchar_x, nchar_y, nchar_max, dist, ...)
}
\arguments{
\item{nchar_x}{Number of characters in token \code{x}}

\item{nchar_y}{Number of characters in token \code{y}}

\item{nchar_max}{\code{max(c(x_nchar, y_nchar))}}

\item{dist}{String distance between tokens x and y}

\item{...}{Additional arguments (not used)}
}
\value{
Logical vector indicating whether tokens \code{x} and \code{y} match, based on their
respective lengths and the string distance between them
}
\description{
Evaluate string distance and token lengths to determine whether two tokens
match
}
