% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{names_ex}
\alias{names_ex}
\title{Example data with proper names from two different sources}
\format{
A data.frame with 6 rows and 2 variables, both of class character:
\describe{
\item{name_source1}{name from source 1}
\item{name_source2}{name from source 2}
}
}
\usage{
names_ex
}
\description{
Example data with proper names from two different sources
}
\keyword{datasets}
