% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmatch-package.R
\docType{package}
\name{nmatch-package}
\alias{nmatch-package}
\title{nmatch: Fuzzy Matching For Proper Names}
\description{
Tools for comparing sets of proper names, accounting for common types of variation in format and style.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/epicentre-msf/nmatch}
  \item Report bugs at \url{https://github.com/epicentre-msf/nmatch/issues}
}

}
\author{
\strong{Maintainer}: Patrick Barks \email{patrick.barks@epicentre.msf.org}

}
\keyword{internal}
