% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmatch_fast_similarity.R
\name{nmatch_fast_similarity}
\alias{nmatch_fast_similarity}
\title{Test version of nmatch_fast that evaluates string similarity scores}
\usage{
nmatch_fast_similarity(
  x,
  y,
  token_split = "[-_[:space:]]+",
  nchar_min = 2L,
  std = name_standardize,
  ...,
  token_freq = NULL
)
}
\arguments{
\item{x, y}{Vectors of proper names to compare. Must be of same length.}

\item{token_split}{Regex pattern to split strings into tokens. Defaults to
\code{"[-_[:space:]]+"}, which splits at each sequence of one more dash,
underscore, or space character.}

\item{nchar_min}{Minimum token size to compare. Defaults to \code{2L}.}

\item{std}{Function to standardize strings during matching. Defaults to
\code{\link{name_standardize}}. Set to \code{NULL} to omit standardization.}

\item{...}{additional arguments passed to \code{std()}}

\item{token_freq}{Optional data frame containing the frequencies of name
tokens within the population of interest. Must have two columns
\itemize{
\item token_std: standardized tokens (using the same function as \code{std})
\item freq: token frequency
}}
}
\value{
Returns a data frame summarizing the match details, including columns:
\itemize{
\item \code{k_x}: number of tokens in \code{x} (excludes tokens smaller than \code{nchar_min})
\item \code{k_y}: number of tokens in \code{y} (excludes tokens smaller than \code{nchar_min})
\item \code{k_align}: number of aligned tokens (i.e. \code{min(k_x, k_y)})
\item \code{sim_total}: summed similarity score across aligned tokens (between 0 and
\code{k_align}, where \code{k_align} indicates a perfect match)
\item \code{freq1}: summed frequency of first pair of aligned tokens (or NA if
argument \code{token_freq} not provided)
\item \code{freq2}: summed frequency of second pair of aligned tokens (or NA if
argument \code{token_freq} not provided)
\item \code{freq3}: summed frequency of third pair of aligned tokens (or NA if
argument \code{token_freq} not provided)
}
}
\description{
Test version of nmatch_fast that evaluates string similarity scores
}
\examples{
names1 <- c(
  "Angela Dorothea Merkel",
  "Emmanuel Jean-Michel Fr\u00e9d\u00e9ric Macron",
  "Mette Frederiksen",
  "Katrin Jakobsd\u00f3ttir",
  "Pedro S\u00e1nchez P\u00e9rez-Castej\u00f3n"
)

names2 <- c(
  "MERKEL, Angela",
  "MACRON, Emmanuel J.-M. F.",
  "FREDERICKSON, Mette",
  "JAKOBSDOTTIR  Kathríne",
  "PEREZ-CASTLEJON, Pedro"
)

# return data frame with match details
nmatch_fast_similarity(names1, names2)

}
