% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmatch_fast_prob.R
\name{nmatch_fast_prob}
\alias{nmatch_fast_prob}
\title{Test version of nmatch_fast that explicitly incorporate false-positive probabilities
for every token pair in the best alignment}
\usage{
nmatch_fast_prob(
  x,
  y,
  token_split = "[-_[:space:]]+",
  nchar_min = 2L,
  std = name_standardize,
  ...,
  token_x,
  dist_x,
  prob_x,
  token_y,
  dist_y,
  prob_y
)
}
\arguments{
\item{x, y}{Vectors of proper names to compare. Must be of same length.}

\item{token_split}{Regex pattern to split strings into tokens. Defaults to
\code{"[-_[:space:]]+"}, which splits at each sequence of one more dash,
underscore, or space character.}

\item{nchar_min}{Minimum token size to compare. Defaults to \code{2L}.}

\item{std}{Function to standardize strings during matching. Defaults to
\code{\link{name_standardize}}. Set to \code{NULL} to omit standardization.}

\item{...}{additional arguments passed to \code{std()}}

\item{token_x}{Vector of standardized tokens for the \code{x}-side lookup table}

\item{dist_x}{Integer vector of OSA distances corresponding to each token in \code{token_x}}

\item{prob_x}{Numeric vector of probabilities corresponding to each token in \code{token_x}}

\item{token_y}{Vector of standardized tokens for the \code{y}-side lookup table}

\item{dist_y}{Integer vector of OSA distances corresponding to each token in \code{token_y}}

\item{prob_y}{Numeric vector of probabilities corresponding to each token in \code{token_y}}
}
\value{
Returns a data frame summarizing the match details, including columns:
\itemize{
\item \code{k_x}: number of tokens in \code{x} (excludes tokens smaller than \code{nchar_min})
\item \code{k_y}: number of tokens in \code{y} (excludes tokens smaller than \code{nchar_min})
\item \code{k_align}: number of aligned tokens (i.e. \code{min(k_x, k_y)})
\item \code{n_match}: number of tokens that match (see \link{match_eval_token} for match logic)
\item \code{dist_total}: summed string distance across aligned tokens
\item \code{p1}: probability of observing a match as good as that between the first aligned token pair
\item \code{p2}: probability of observing a match as good as that between the second aligned token pair
\item \code{p3}: probability of observing a match as good as that between the third aligned token pair
\item \code{p_match}: overall match probability
}
}
\description{
Test version of nmatch_fast that explicitly incorporate false-positive probabilities
for every token pair in the best alignment
}
